Remove unlock retry
Unlocking is now a simple `rm -f`. If that fails things are broken on a
larger scale.

This removes a small race condition: between unlocking and checking the
lock state the same user might have requested the lock again and been
successful. The retry would incorrectly free that new lock.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>