remote: configure ssh not to produce \r\n
The default configuration produced \r\n for newlines (with -t, which we
used), leaving \r characters in the output at unexpected places. This
configures ssh not to do that, and removes the `tr -d '\r'` sprinkled
through the code.

We only do this for RemoteCommand, where we expect one-line output,
because the terminal does actually seem to expect \r characters for
multi-line things.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>