Add timeout to locks
A new argument, -T, to the locking commands adds a timeout
in seconds to the lock.  The default is 'forever' to preserve
current behaviour (this will probably change in a future commit)

This works by storing the timeout in the lockfile.  When a new lock
is taken for the same system if more time has elapsed than the timeout,
the lock will be broken, so the new lock can be held.

Signed-off-by: Peter Chubb <peter.chubb@unsw.edu.au>